Allow enable or diable unauthorized clients persistence

+++ application/src/main/java/org/thingsboard/mqtt/broker/service/auth/unauthorized/UnauthorizedClientManagerImpl.java
@@ -20,13 +20,14 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.thingsboard.mqtt.broker.actors.client.messages.SessionInitMsg;
 import org.thingsboard.mqtt.broker.actors.client.state.ClientActorState;
+import org.thingsboard.mqtt.broker.common.data.BrokerConstants;
 import org.thingsboard.mqtt.broker.common.data.UnauthorizedClient;
 import org.thingsboard.mqtt.broker.common.data.util.BytesUtil;
 import org.thingsboard.mqtt.broker.common.util.DonAsynchron;
+import org.thingsboard.mqtt.broker.config.UnauthorizedClientsProperties;
 import org.thingsboard.mqtt.broker.dao.client.unauthorized.UnauthorizedClientService;
 import org.thingsboard.mqtt.broker.service.auth.enhanced.EnhancedAuthContinueResponse;
 import org.thingsboard.mqtt.broker.service.auth.enhanced.EnhancedAuthFinalResponse;
-import org.thingsboard.mqtt.broker.service.auth.providers.AuthResponse;
 import org.thingsboard.mqtt.broker.session.ClientSessionCtx;
 
 @Slf4j
@@ -35,11 +36,7 @@ import org.thingsboard.mqtt.broker.session.ClientSessionCtx;
 public class UnauthorizedClientManagerImpl implements UnauthorizedClientManager {
 
     private final UnauthorizedClientService unauthorizedClientService;
-
-    @Override
-    public void persistClientUnauthorized(ClientActorState state, SessionInitMsg sessionInitMsg, AuthResponse authResponse) {
-        persistClientUnauthorized(state, sessionInitMsg, authResponse.getReason());
-    }
+    private final UnauthorizedClientsProperties props;
 
     @Override
     public void persistClientUnauthorized(ClientActorState state, SessionInitMsg sessionInitMsg, String reason) {
@@ -62,6 +59,9 @@ public class UnauthorizedClientManagerImpl implements UnauthorizedClientManager 
     @Override
     public void persistClientUnauthorized(ClientActorState state, ClientSessionCtx clientSessionCtx,
                                           String username, boolean passwordProvided, String reason) {
+        if (!props.isEnabled()) {
+            return;
+        }
         UnauthorizedClient unauthorizedClient = UnauthorizedClient.builder()
                 .clientId(state.getClientId())
                 .ipAddress(BytesUtil.toHostAddress(clientSessionCtx.getAddressBytes()))
@@ -75,7 +75,19 @@ public class UnauthorizedClientManagerImpl implements UnauthorizedClientManager 
     }
 
     @Override
-    public void persistClientUnauthorized(UnauthorizedClient unauthorizedClient) {
+    public void persistClientUnauthorized(String clientId, String reason) {
+        if (!props.isEnabled()) {
+            return;
+        }
+        UnauthorizedClient unauthorizedClient = UnauthorizedClient.builder()
+                .clientId(clientId)
+                .ipAddress(BrokerConstants.UNKNOWN) // can be improved later
+                .ts(System.currentTimeMillis())
+                .username(BrokerConstants.UNKNOWN)
+                .passwordProvided(false)
+                .tlsUsed(true)
+                .reason(reason)
+                .build();
         persist(unauthorizedClient);
     }
 
@@ -87,9 +99,10 @@ public class UnauthorizedClientManagerImpl implements UnauthorizedClientManager 
 
     @Override
     public void removeClientUnauthorized(ClientActorState state) {
-        UnauthorizedClient unauthorizedClient = UnauthorizedClient.builder()
-                .clientId(state.getClientId())
-                .build();
+        if (!props.isEnabled()) {
+            return;
+        }
+        UnauthorizedClient unauthorizedClient = UnauthorizedClient.builder().clientId(state.getClientId()).build();
         DonAsynchron.withCallback(unauthorizedClientService.remove(unauthorizedClient),
                 v -> log.debug("[{}] Unauthorized Client removed successfully!", state.getClientId()),
                 throwable -> log.warn("[{}] Failed to removed unauthorized client!", state.getClientId(), throwable));
